move Query, Prompt, Context to data_structures.rag
