+++ src/fed_rag/data_structures/__init__.py
@@ -11,9 +11,8 @@ from .evals import (
     BenchmarkExample,
     BenchmarkResult,
 )
-from .generator import Context, Prompt, Query
 from .knowledge_node import KnowledgeNode, NodeContent, NodeType
-from .rag import RAGConfig, RAGResponse, SourceNode
+from .rag import Context, Prompt, Query, RAGConfig, RAGResponse, SourceNode
 from .results import TestResult, TrainResult
 
 __all__ = [
@@ -25,10 +24,6 @@ __all__ = [
     "BenchmarkExample",
     "BenchmarkResult",
     "BenchmarkEvaluatedExample",
-    # generator
-    "Query",
-    "Context",
-    "Prompt",
     # results
     "TrainResult",
     "TestResult",
@@ -40,4 +35,7 @@ __all__ = [
     "RAGConfig",
     "RAGResponse",
     "SourceNode",
+    "Query",
+    "Context",
+    "Prompt",
 ]
